requirements.txt
appnope==0.1.2
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
cffi==1.14.5
click==7.1.2
colorama==0.4.4
decorator==5.0.7
dnspython==2.1.0
email-validator==1.1.2
Faker==8.1.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.0.0
idna==3.1
ipython==7.22.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
MarkupSafe==1.1.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.1
python-dateutil==2.8.1
simplegeneric==0.8.1
six==1.15.0
SQLAlchemy==1.4.7
text-unidecode==1.3
traitlets==5.0.5
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
